requirements.txt
Flask==0.11.1
requests==2.10.0
google-cloud-language==0.22.2
google-cloud-vision==0.22.0
